requirements.txt
Django==4.0.1
discord.py==1.7.3
